requirements.txt
flask~=2.3.3
flask-cors~=6.0.1
gspread~=6.2.1
google-auth
gunicorn~=23.0.0
python-dotenv~=1.1.1

pip~=25.2
attrs~=25.3.0
PySocks~=1.7.1
pillow~=11.3.0
rsa~=4.9.1
pyasn1~=0.6.1
typing_extensions~=4.14.1
MarkupSafe~=3.0.2
Werkzeug~=3.1.3
click~=8.2.1
Jinja2~=3.1.6
blinker~=1.9.0
itsdangerous~=2.2.0
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.8.3
urllib3~=2.5.0
packaging~=25.0
requests~=2.32.4
cffi~=1.17.1
google-auth-oauthlib~=1.2.2
greenlet~=3.2.4
oauthlib~=3.3.1
idna~=3.10
charset-normalizer~=3.4.3
requests-oauthlib~=2.0.0